feat: cache dicussions
